[#24] refactor: data 패키지를 도메인별로 구성 (member)
기술 타입(entity/enums/repository) 분리 → 도메인 묶음으로 전환.
- data.member: Member, MemberRepository
- data.member.enums: Position, MemberStatus, ProgressStatus
- 테스트도 data.member로 이동, 패키지/임포트 정리
- 향후 attendance 등도 data.<도메인> 동일 패턴 (저널 기록)
