+++ docs/agent-journal/2026-06-15-issue-24.md
@@ -101,4 +101,6 @@
 | 상태 `status` | `MemberStatus`(해외/장결자/군대) | **정상 = null** |
 | 행삶 `action`·양육 `training` | `ProgressStatus`(미진행/진행중/완료) | 미상 = null |
 
+**패키지 구성 — 도메인별:** 기술 타입(entity/enums) 대신 도메인으로 묶는다. 현재는 member 도메인뿐이라 `data.member`(엔티티·리포지토리) + `data.member.enums`(enum). 향후 attendance·representative-prayer 데이터도 `data.<도메인>`으로 동일 패턴. (`enum`은 예약어라 폴더명은 `enums`)
+
 **남은 결정:** 생년월일 파싱(연도 보정) 로직은 엑셀→엔티티 매퍼(#21~#23)에 구현. 로그인 ID는 `Member`가 아닌 `account.User`(feature/03)에 두고, 연결이 필요해지면 스냅샷 id가 아닌 안정 키로 소프트 링크.
