+++ src/main/java/com/saeanyang/management/data/member/MemberRepository.java
@@ -1,6 +1,5 @@
-package com.saeanyang.management.data.repository;
+package com.saeanyang.management.data.member;
 
-import com.saeanyang.management.data.entity.Member;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
